Refactor tests and address PR comments
- Created `TestUtils.kt` to share `TestListenableFuture` and `FailingListenableFuture`.
- Refactored `Jp2kDecoderTest`, `Jp2kDecoderAsyncCoverageTest`, and `Jp2kOverloadsTest` to use the shared utilities.
- Improved code style by using float literals (`0.0f`).
- Added explicit verification for logging calls in tests.
